Remove trino dependencies (#523)

+++ plugin/writer/hdfswriter/src/main/java/com/wgzhao/addax/plugin/writer/hdfswriter/HdfsHelper.java
@@ -577,7 +577,7 @@ public class HdfsHelper
 
         try (ParquetWriter<GenericRecord> writer = AvroParquetWriter
                 .<GenericRecord>builder(path)
-                .withRowGroupSize((long) ParquetWriter.DEFAULT_BLOCK_SIZE)
+                .withRowGroupSize(ParquetWriter.DEFAULT_BLOCK_SIZE)
                 .withPageSize(ParquetWriter.DEFAULT_PAGE_SIZE)
                 .withSchema(schema)
                 .withConf(hadoopConf)
